requirements/local.txt
-r base.txt

Werkzeug==1.0.1 # https://github.com/pallets/werkzeug
ipdb==0.13.3  # https://github.com/gotcha/ipdb
psycopg2==2.8.5 --no-binary psycopg2  # https://github.com/psycopg/psycopg2

# Testing
# ------------------------------------------------------------------------------
mypy==0.770  # https://github.com/python/mypy
django-stubs==1.5.0  # https://github.com/typeddjango/django-stubs
pytest==6.0.1  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.4  # https://github.com/Frozenball/pytest-sugar

# Documentation
# ------------------------------------------------------------------------------
sphinx==3.2.1  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild==0.7.1 # https://github.com/GaretJax/sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
flake8==3.8.3  # https://github.com/PyCQA/flake8
flake8-isort==4.0.0  # https://github.com/gforcada/flake8-isort
coverage==5.2.1  # https://github.com/nedbat/coveragepy
black==19.10b0  # https://github.com/ambv/black
pylint-django==2.3.0  # https://github.com/PyCQA/pylint-django
pylint-celery==0.3  # https://github.com/PyCQA/pylint-celery
pre-commit==2.7.1  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==3.0.1  # https://github.com/FactoryBoy/factory_boy

django-extensions==3.0.5  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==1.8.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==3.9.0  # https://github.com/pytest-dev/pytest-django
django-phonenumber-field[phonenumbers]==5.0.0
